add pundit, add pizza_types policy

+++ app/services/order_service.rb
@@ -21,6 +21,6 @@ class OrderService
   end
 
   def _discount
-    Order::DISCOUNTS[Order::DISCOUNT_CONDITIONS.key(Order::DISCOUNT_CONDITIONS.values.find_all { |v| v <= @order.cost.to_i }.last)]
+    Order::DISCOUNTS[Order::DISCOUNT_CONDITIONS.key(Order::DISCOUNT_CONDITIONS.values.find_all { |v| v <= @order.cost.to_f }.last)]
   end
 end
